third_party/blink/web_tests/platform/linux/external/wpt/html/browsers/windows/clear-window-name.https-expected.txt

This is a testharness.js-based test.
PASS Window.name is not reset when there is an opener around
PASS Window.name is not reset when there is an opener around (cross-origin)
PASS Window.name is not reset at the first navigation away from initial about:blank with noopener
FAIL Window.name is reset at the first cross-origin navigation with noopener assert_equals: expected "" but got "73d3b9ef-1a2d-45ce-8b97-575a37d635d3"
PASS Window.name is not reset at the first navigation away from initial about:blank with window.opener set to null
PASS Window.name is not reset at the same-origin navigation with window.opener set to null
FAIL Window.name is reset at the first first cross-origin navigation with window.opener set to null assert_equals: expected "" but got "f8518422-a53e-4149-b4aa-e592a7c4d5d4"
PASS Window.name is set by the window
FAIL Window.name is reset at the first cross-origin navigation assert_equals: expected "" but got "f5f87800-3a80-4fcb-a2e0-72a3f42bbf14"
Harness: the test ran to completion.
